fix: 增加之前mvu提供的 event
